Add service group
Add the service group that can register and run component parts of a
service and run them.

If any of the functions registered return an error, the entire service
is stopped.

The service exposes Liveness and Readiness probes to report if the setup
and running of the service has completed.